feat: seperate quiz function and make it to use mongodb
